HomeWork-7 is done...
